Web-automation (TESTING)
Because of a few issues, Selenium may be replaced with another method later. Grabbing an existing browser session with Selenium doesn't seem possible without a workaround, so this script is likely to be captcha'd or rate limited. On top of that, selenium doesn't seem to be able to grab the chrome profile even when directed. Making auto sign-in impossible.
So far this small draft script does the following:
- Opens a browser
- Signs into Grammarly
- Opens the "create new document" page.